URL injection protection ! ^^

+++ admin/includes/view_all_comments.php
@@ -74,14 +74,22 @@
           }
 
           if(isset($_GET['delete'])){
-            $the_comment_id = $_GET['delete'];
 
-            $query ="DELETE FROM comments WHERE comment_id = $the_comment_id ";
-            $delete_query = mysqli_query($connection, $query);
+                if(isset($_SESSION['role'])) {
 
-            header('Location: comments.php');
+                  if($_SESSION['role'] == 'admin') {
+                  $the_comment_id = $_GET['delete'];
 
-          }
+                  $query ="DELETE FROM comments WHERE comment_id = $the_comment_id ";
+                  $delete_query = mysqli_query($connection, $query);
+
+                  header('Location: comments.php');
+
+                }
+
+              }
+              
+            }
 
 
         ?>
