+++ admin/includes/view_all_posts.php
@@ -186,12 +186,21 @@ if(isset($_POST['checkBoxArray'])){
             }
 
             if(isset($_GET['delete'])){
-              $the_post_id = $_GET['delete'];
 
-              $query ="DELETE FROM posts WHERE post_id = $the_post_id ";
-              $delete_query = mysqli_query($connection, $query);
+              if(isset($_SESSION['role'])) {
 
-              header('Location: posts.php');
+                if($_SESSION['role'] == 'admin') {
+
+                  $the_post_id = $_GET['delete'];
+
+                  $query ="DELETE FROM posts WHERE post_id = $the_post_id ";
+                  $delete_query = mysqli_query($connection, $query);
+
+                  header('Location: posts.php');
+                  
+                }
+
+              }
 
             }
